Implement stubs + wire into brain: AlphaDiscoveryEngine (GA), FractalAgent, CorrelationMonitor, CapitalFlowAgent, BayesianRegimeAgent; fix fractal int-cast NaN, shape mismatch in GA scoring, correlation risk_multiplier clamp, option squeeze bool; 42 new tests all green

+++ src/fractal_agent.py
@@ -39,7 +39,7 @@ class FractalAgent:
                     for i in range(1, max_i + 1):
                         Lmk += abs(data_arr[m + i * k] - data_arr[m + (i - 1) * k])
 
-                    Lmk = int(Lmk * (N - 1) / (max_i * k) / k)
+                    Lmk = Lmk * (N - 1) / (max_i * k) / k
                     Lk.append(Lmk)
 
             if Lk:
